moving vertices and edges is now performed relative to the model, and a preset for selecting the relative movement has been added.

+++ WindowsFormsApp3/3D/Model/Grid/Gizmo3D.cs
@@ -277,7 +277,7 @@ namespace WindowsFormsApp3
 
         public static Vector3 ToOpenTK(Assimp.Vector3D v)
         {
-            return new Vector3((float)v.X, (float)v.Y, (float)v.Z);
+            return new Vector3(v.X, v.Y, v.Z);
         }
 
         public static Assimp.Vector3D ToAssimp(Vector3 v)
